sub 0.3.30 LS.2 삭제 제대로 안되던 오류 수정

+++ BuildingBuilder.h
@@ -52,7 +52,7 @@ public:
 
 	void updateTerrainNodeData(GameObject* building, bool collision);
 
-	void build(Vector2 position, float angle, int type, int index, GameObject* landmark = nullptr);
+	void build(Vector2 position, double angle, int type, int index, GameObject* landmark = nullptr);
 
 	void build(Vector3 position);
